+++ src/bootstrap/clean.rs
@@ -9,8 +9,7 @@ use std::fs;
 use std::io::{self, ErrorKind};
 use std::path::Path;
 
-use build_helper::t;
-
+use crate::build_helper::t;
 use crate::Build;
 
 pub fn clean(build: &Build, all: bool) {
