+++ src/bootstrap/metadata.rs
@@ -1,9 +1,9 @@
 use std::path::PathBuf;
 use std::process::Command;
 
-use build_helper::output;
 use serde::Deserialize;
 
+use crate::build_helper::output;
 use crate::cache::INTERNER;
 use crate::{Build, Crate};
