+++ src/bootstrap/native.rs
@@ -16,13 +16,12 @@ use std::io;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
-use build_helper::{output, t};
-
+use crate::build_helper::up_to_date;
+use crate::build_helper::{output, t};
 use crate::builder::{Builder, RunConfig, ShouldRun, Step};
 use crate::config::TargetSelection;
 use crate::util::{self, exe};
 use crate::{CLang, GitRepo};
-use build_helper::up_to_date;
 
 pub struct Meta {
     stamp: HashStamp,
