+++ src/bootstrap/toolstate.rs
@@ -1,5 +1,5 @@
+use crate::build_helper::t;
 use crate::builder::{Builder, RunConfig, ShouldRun, Step};
-use build_helper::t;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::env;
