--- src/build_helper/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "build_helper"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "lib.rs"
